pyproject.toml
[project]
name = "medical-consultation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "whitenoise>=6.9.0",
]
